# docs/api/openapi-bundled.yaml
openapi: 3.1.0
info:
  title: AuthFramework REST API
  description: |
    Comprehensive REST API for the AuthFramework authentication and authorization library.

    This API provides complete access to AuthFramework's authentication, authorization,
    user management, multi-factor authentication, OAuth 2.0, OIDC, RBAC, and administrative 
    features through standardized HTTP endpoints.

    ## Version Strategy

    - **Versioned endpoints** (`/api/v1/*`): All functional API endpoints are versioned
      for stability and backward compatibility. Version changes follow semantic versioning.

    - **Unversioned endpoints**: Infrastructure endpoints (`/health`, `/metrics`, `/readiness`, 
      `/liveness`) and OIDC discovery (`/.well-known/*`) remain unversioned for operational 
      stability and spec compliance.

    ## Authentication

    Most endpoints require authentication via Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```

    Obtain tokens via:
    - `/api/v1/auth/login` - Username/password authentication
    - `/api/v1/oauth/token` - OAuth 2.0 token endpoint
    - `/api/v1/auth/refresh` - Refresh token flow

    ## Rate Limiting

    API endpoints are rate-limited to prevent abuse:
    - Authentication endpoints: 5 requests per minute per IP
    - Standard endpoints: 100 requests per minute per user
    - Admin endpoints: 50 requests per minute per admin user

...

    ## Response Format

...
    ```json
    {
...
    }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/oauth/userinfo:
    get:
      tags:
        - OAuth
        - OIDC
      summary: OpenID Connect UserInfo
      description: |
        OIDC UserInfo endpoint (OIDC Core 5.3). Returns claims about the authenticated user.
        Requires valid access token with 'openid' scope.
      operationId: userinfo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Invalid or missing access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: invalid_token
                  error_description:
                    type: string
                    example: The access token is invalid or expired
        '403':
          description: Insufficient scope (openid scope required)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: insufficient_scope
                  error_description:
                    type: string
                    example: The request requires 'openid' scope
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-access-token>
        ```
...
  schemas:
...
    UserInfoResponse:
      type: object
      description: OpenID Connect UserInfo response (OIDC Core 5.3)
      properties:
        sub:
          type: string
          description: Subject identifier (user ID)
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Full name
          example: John Doe
        given_name:
          type: string
          description: First name
          example: John
        family_name:
          type: string
          description: Last name
          example: Doe
        email:
          type: string
          format: email
          description: Email address
          example: user@example.com
        email_verified:
          type: boolean
          description: Whether email has been verified
          example: true
        phone_number:
          type: string
          description: Phone number
          example: '+1234567890'
        phone_number_verified:
          type: boolean
          description: Whether phone number has been verified
          example: false
        picture:
          type: string
          format: uri
          description: Profile picture URL
          example: https://example.com/avatars/user123.jpg
        locale:
          type: string
          description: User's locale
          example: en-US
        zoneinfo:
          type: string
          description: User's timezone
          example: America/New_York
        updated_at:
          type: integer
          description: Last profile update timestamp (Unix epoch)
          example: 1642686645
      required:
        - sub
...